Extract WorkingSet to separate file and finalize mod.rs
- Created scratchpad/working_set.rs with WorkingSet and all impls
- Moved tests to scratchpad/mod.rs tests module
- Updated scratchpad/mod.rs to be a clean module with re-exports
- Added comprehensive module documentation
- Removed all unused imports
- Code compiles with no warnings
